create customer api

+++ src/main/java/com/witnes/CustomerRepository.java
@@ -2,5 +2,5 @@ package com.witnes;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CustomerRepository extends JpaRepository<Customer,Integer> {
+public interface CustomerRepository extends JpaRepository<Customer, Long> { // Change Integer to Long
 }
